Split sample mutation helpers
